Refactoring - Move create.html and code from dashboard views and urls to learning resource
