+++ code_resource_center/urls.py
@@ -23,5 +23,6 @@ urlpatterns = [
     path('', accounts.views.login, name='login'),
     path('dashboard/', views.dashboard, name='dashboard'),
     path('accounts/', include('accounts.urls')),
-    path('dashboard/', include('dashboard.urls')),
+    path('learningresource/', include('learningresource.urls')),
+
 ]
