+++ dashboard/urls.py
@@ -1,6 +1,1 @@
 from django.urls import path, include
-from . import views
-
-urlpatterns = [
-    path('create', views.create, name ='create')
-]
